+++ tests/test_cloudflare.py
@@ -3,12 +3,9 @@ from unittest.mock import MagicMock, patch
 
 import CloudFlare
 import pytest
-from cloudflare_companion import (
-    CloudFlareZones,
-    DomainsModel,
-    Settings,
-    Singleton,
-)
+from cloudflare_companion import CloudFlareMapper, Settings
+from internal._singleton import Singleton
+from settings import DomainsModel
 
 
 @pytest.fixture
@@ -51,7 +48,7 @@ def mock_settings(mock_domain_infos):
 
 @pytest.fixture(scope="module", autouse=True)
 def patch_singleton():
-    side_effect = super(Singleton, CloudFlareZones).__call__
+    side_effect = super(Singleton, CloudFlareMapper).__call__
     with patch.object(Singleton, "__call__", side_effect=side_effect):
         yield
 
@@ -60,7 +57,7 @@ def patch_singleton():
 async def test_update_zones_target_domain_match(
     mock_cloudflare, mock_settings, mock_domain_infos, mock_logger
 ):
-    cf_zones = CloudFlareZones(mock_logger, settings=mock_settings, client=mock_cloudflare)
+    cf_zones = CloudFlareMapper(mock_logger, settings=mock_settings, client=mock_cloudflare)
     result = await cf_zones.update_zones("target.example.com", mock_domain_infos)
     assert result is True
     mock_cloudflare.zones.dns_records.get.assert_not_called()
@@ -71,7 +68,7 @@ async def test_update_zones_excluded_domain(
     mock_cloudflare, mock_settings, mock_domain_infos, mock_logger
 ):
     mock_domain_infos[0].excluded_sub_domains = ["sub"]
-    cf_zones = CloudFlareZones(mock_logger, settings=mock_settings, client=mock_cloudflare)
+    cf_zones = CloudFlareMapper(mock_logger, settings=mock_settings, client=mock_cloudflare)
     result = await cf_zones.update_zones("sub.example.com", mock_domain_infos)
     assert result is True
     mock_cloudflare.zones.dns_records.get.assert_not_called()
@@ -82,7 +79,7 @@ async def test_update_zones_create_new_record(
     mock_cloudflare, mock_settings, mock_domain_infos, mock_logger
 ):
     mock_cloudflare.zones.dns_records.get.return_value = []
-    cf_zones = CloudFlareZones(mock_logger, settings=mock_settings, client=mock_cloudflare)
+    cf_zones = CloudFlareMapper(mock_logger, settings=mock_settings, client=mock_cloudflare)
     result = await cf_zones.update_zones("new.example.com", mock_domain_infos)
     assert result is True
     mock_cloudflare.zones.dns_records.post.assert_called_once()
@@ -93,7 +90,7 @@ async def test_update_zones_update_existing_record(
     mock_cloudflare, mock_settings, mock_domain_infos, mock_logger
 ):
     mock_cloudflare.zones.dns_records.get.return_value = [{"id": "record_id"}]
-    cf_zones = CloudFlareZones(mock_logger, settings=mock_settings, client=mock_cloudflare)
+    cf_zones = CloudFlareMapper(mock_logger, settings=mock_settings, client=mock_cloudflare)
     result = await cf_zones.update_zones("existing.example.com", mock_domain_infos)
     assert result is True
     mock_cloudflare.zones.dns_records.put.assert_called_once()
@@ -108,7 +105,7 @@ async def test_update_zones_rate_limit_retry(
         CloudFlare.exceptions.CloudFlareAPIError(-1, "Rate limited"),
         [],
     ]
-    cf_zones = CloudFlareZones(mock_logger, settings=mock_settings, client=mock_cloudflare)
+    cf_zones = CloudFlareMapper(mock_logger, settings=mock_settings, client=mock_cloudflare)
     with patch("asyncio.sleep", return_value=None):
         result = await cf_zones.update_zones("rate_limited.example.com", mock_domain_infos)
     assert result is True
@@ -119,7 +116,7 @@ async def test_update_zones_rate_limit_retry(
 async def test_update_zones_dry_run(mock_cloudflare, mock_settings, mock_domain_infos, mock_logger):
     mock_settings.dry_run = True
     mock_cloudflare.zones.dns_records.get.return_value = []
-    cf_zones = CloudFlareZones(mock_logger, settings=mock_settings, client=mock_cloudflare)
+    cf_zones = CloudFlareMapper(mock_logger, settings=mock_settings, client=mock_cloudflare)
     result = await cf_zones.update_zones("dryrun.example.com", mock_domain_infos)
     assert result is True
     mock_cloudflare.zones.dns_records.post.assert_not_called()
